refactor(server): isolate filesystem durability helpers
